Minor bug fix and update
Move two chown calls out of if statements. The if statements are
for calling makedirs if directory doesn't exist. Always want to
set owner to current tester user to avoid issues with file
permissions.

Also updated exec_string for make to include a new variable
called CPATH which has the directory make will run in. This is
need for java makefiles to build CLASSPATH for compiling.