Adds automated grading to API.
Submissions are automatically graded after submission now. Any tests
that error out will result in a 'max' of -1 points.

For each test, the weights of all subtests are summed and the weights
of all passing subtests are summed. If the test itself does not have
an assigned point value, then these values are added directly to the
running totals for the submission.

If the test does have an overall point value assigned, then the weights
are scaled using the formula:
   passing_weight/total_weight * test_point_value
and then the scaled value and the test point value are added to the
running totals for the submission.